Fix dw2pdf multipage struct output
When the dw2pdf plugin is used to export multiple pages in a
single document, all the pages are rendered in the same context
of a single page. In order to include struct data in the PDF,
let's skip the id related checks and allow multiple pages rendered
in a single instance of syntax_plugin_struct_output.

Signed-off-by: Frieder Schrempf <dev@fris.de>